Add support for jemalloc
Typically I think Jemalloc is normally only used for multithreaded programs, to prevent some performance issues that can happen with other allocators in that situation. However, this also fixes (or at least masks) a weird, inexplicable bug that was happening with Glibc default malloc on long (>1000 step) programs and causing a segfault crash.